+++ doc/source/changes.rst
@@ -17,8 +17,10 @@ Versions
    * Add registration frontend for new clients.
    * Add security support for endpoints / controllers.
    * Add Fantastico identity provider.
-   * Fix a bug in ROA APIs (on some python 3 compilers get_item operations were failing with resource not found).
+   * Fix a bug in ROA APIs routes mapping.
    * Added support for MVC Controllers into custom packages (not residing in components root folder).
+   * Added ROA resources dependent on user (integrated with OAuth2 access tokens).
+   * Added ROA resources OAuth2 authorization.
 
 * v0.5.1 (`Provide feedback <https://docs.google.com/forms/d/1tKBopU3lfDB_w8F4h7Rl1Rn4uydAJq-nha09L_ptJck/edit?usp=sharing>`_)