+++ doc/source/changes.rst
@@ -11,6 +11,10 @@ so do not hesitate to get in touch with us (for support, feature requests, sugge
 Versions
 --------
 
+* v0.7.0 (Work in progress)
+
+   * Added support for cross origin requests on ROA resources registry (/roa/resources)
+
 * v0.6.0 (`Provide feedback <https://docs.google.com/forms/d/1tKBopU3lfDB_w8F4h7Rl1Rn4uydAJq-nha09L_ptJck/edit?usp=sharing>`_)
 
    * Added implicit grant type implementation.